add modalContent in useModalStore, fixed Modal.jsx - now is just container
